pyproject.toml
[project]
name = "docuchat"
version = "0.1.0"
description = "RAG-powered document assistant using LangChain + Groq AI"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "hatchling>=1.29.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.17",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
